requirements.txt
streamlit
yfinance
ta
python-dotenv
plotly
requests
numpy
openai
google-auth
google-auth-oauthlib
google-api-python-client